+++ src/Survey/Plant/config/index.tsx
@@ -2,14 +2,11 @@
 import {
   peopleOutline,
   businessOutline,
-  pencilOutline,
   eyeOffOutline,
   timeOutline,
 } from 'ionicons/icons';
-import { Trans as T } from 'react-i18next';
 import { object, array, string } from 'zod';
 import {
-  InfoButton,
   dateFormatISO,
   type OccurrenceData,
   type SampleData,
@@ -24,7 +21,6 @@ import { IonIcon } from '@ionic/react';
 import { groupsReverse as groups } from 'common/data/informalGroups';
 import VCs from 'common/data/vice_counties.data.json';
 import gridAlertService from 'common/helpers/gridAlertService';
-import numberIcon from 'common/images/number.svg';
 import Sample from 'common/models/sample';
 import appModel from 'models/app';
 import Occurrence, { MachineInvolvement } from 'models/occurrence';
@@ -41,14 +37,24 @@ import {
   plantStageAttr,
 } from 'Survey/common/config';
 import bryophytesSurvey from './bryophytes';
-import plantOccIdentifiersAttr, { altitudeAttr } from './common';
+import plantOccIdentifiersAttr, {
+  abundanceAttr,
+  abundanceSchema,
+  altitudeAttr,
+  statusAttr,
+} from './common';
 
 export {
   commentAttr,
   dateAttr,
   childGeolocationAttr,
 } from 'Survey/common/config';
-export { altitudeAttr, default as plantOccIdentifiersAttr } from './common';
+export {
+  abundanceAttr,
+  altitudeAttr,
+  statusAttr,
+  default as plantOccIdentifiersAttr,
+} from './common';
 
 const plantLocationAttr = {
   ...locationAttr,
@@ -158,37 +164,6 @@ const plantSmpLocationAttr = {
   },
 } as const;
 
-export const abundanceAttr = {
-  id: 'occAttr:610',
-  title: 'Abundance',
-  prefix: <IonIcon src={numberIcon} className="size-6" />,
-  type: 'textInput',
-  container: 'page',
-  description: (
-    <T>
-      Abundance (DAFOR, LA, LF or count).
-      <InfoButton label="READ MORE" header="Info" color="tertiary">
-        <p>
-          DAFOR refers to a subjective abundance scale comprising the following
-          ordered terms: <b>D</b>ominant / <b>A</b>
-          bundant / <b>F</b>requent / <b>O</b>ccasional / <b>R</b>
-          are. The prefix "Locally" can also be used with the Abundant and
-          Frequent classes (e.g. LA = Locally Abundant).
-        </p>
-        <p>
-          Assessed abundance should either relate to the scale of the survey
-          (e.g. 1 or 2 km grid squares), or be clearly qualified in the record
-          comments field.
-        </p>
-      </InfoButton>
-    </T>
-  ),
-  validation: { pattern: '^(\\d+|[DAFORdafor]|[Ll][Aa]|[Ll][Ff])$' },
-  onChange: (val, op, { record }) => {
-    record[abundanceAttr.id] = val.toUpperCase();
-  },
-} as const satisfies TextInputConf;
-
 // remove after migration is complete
 /** @deprecated */
 export const statusAttrOld = {
@@ -208,41 +183,6 @@ export const statusAttrOld = {
     ],
   },
 };
-
-export const statusAttr = {
-  id: 'occAttr:507',
-  title: 'Status',
-  prefix: <IonIcon src={pencilOutline} className="size-6" />,
-  type: 'choiceInput',
-  appearance: 'button',
-  choices: [
-    { title: 'Not Recorded', dataName: '' },
-    { title: 'Native - origin unknown', dataName: '17548' },
-    { title: 'Introduced - accidental', dataName: '17549' },
-    { title: 'Introduced - accidental - regenerating', dataName: '17550' },
-    { title: 'Introduced - intentional', dataName: '17551' },
-    { title: 'Introduced - intentional - regenerating', dataName: '17552' },
-    { title: 'Introduced - origin unknown', dataName: '17553' },
-    {
-      title: 'Introduced - origin unknown - regenerating',
-      dataName: '17554',
-    },
-
-    // hidden choices for legacy records
-    { title: 'Native', dataName: '5709', className: 'hidden' },
-    { title: 'Unknown', dataName: '5710', className: 'hidden' },
-    { title: 'Introduced', dataName: '6775', className: 'hidden' },
-    { title: 'Introduced - planted', dataName: '5711', className: 'hidden' },
-    { title: 'Introduced - surviving', dataName: '10662', className: 'hidden' },
-    { title: 'Introduced - casual', dataName: '10663', className: 'hidden' },
-    {
-      title: 'Introduced - established',
-      dataName: '5712',
-      className: 'hidden',
-    },
-    { title: 'Introduced - invasive', dataName: '5713', className: 'hidden' },
-  ],
-} as const satisfies ChoiceInputConf;
 
 export const plantSensitivityPrecisionAttr = {
   id: 'sensitivityPrecision',
@@ -347,6 +287,7 @@ const survey = {
       verify: (values: any) =>
         object({
           taxon: object({}, { error: 'Species is missing.' }).nullable(),
+          [abundanceAttr.id]: abundanceSchema,
         }).safeParse(values).error,
 
       modifySubmission(submission: any, occ: Occurrence) {